Reorganize into agents folder
